requirements.txt
fastapi==0.101.1
uvicorn==0.18.3
gunicorn==20.1.0
psycopg2-binary==2.9.9
qrcode==7.4.2
jinja2==3.1.2   
python-dotenv==1.0.0
python-multipart==0.0.6
boto3==1.28.1
fpdf==1.7.2
python-jose==3.3.0
passlib==1.7.4
pyzbar==0.1.9
pillow==10.2.0
pillow_heif==0.15.0
twilio
apscheduler
pydantic-settings==2.2.1
aiofiles
pydantic==2.6.4
